+++ miscDocumentation/CHANGELOG.md
@@ -1,5 +1,12 @@
 # VOP Release Tracking
 
+## v0.6.3 - (YYYMMDD)
+- Added Noise floor measurement
+- Added Noise Crush input
+- Modified version numbering in the GUI title so it only tracks v#.# instead of v#.#.# . I'll still update the third digit in the header of the GUI though. 
+- Added update to the preview to show what pixels are measured with the noise measurement function.
+- Added Hot Pixel Fixer that measures hot pixels and replaces those pixels in post. 
+
 ## v0.6.2 - (20260406)
 
 ### Added